feat(use-position): separate use-position hook
Separate the `use-position` hook for external usage.
